Re-organize core/ directory
