+++ packages/core/src/express/index.ts
@@ -1,2 +1,1 @@
-export { IController, IAppController } from './app.controller.interface';
 export { createApp, OPENAPI_SERVICE_ID } from './create-app';
